json handles were moved to a new package in #10202 this was unecessary after refactoring, so this moves them back to their original location in package structs
